requirements.txt
asgiref==3.2.3
astroid==2.4.1
blessings==1.7
bpython==0.19
certifi==2019.11.28
chardet==3.0.4
curtsies==0.3.1
Django==3.0.6
djangorestframework==3.11.0
greenlet==0.4.15
httplib2==0.17.3
idna==2.8
isort==4.3.21
jsonpath==0.82
lazy-object-proxy==1.4.3
ldap3==2.7
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.4
pinyin==0.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
PyMySQL==0.9.3
python-ldap==3.2.0
pytz==2019.3
pyvmomi==6.7.3
redis==3.4.1
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.9
wrapt==1.11.2
